pyproject.toml
[project]
name = "parasito"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "pillow>=10.0.0",
]
